+++ App/db/dbsql.js
@@ -11,9 +11,7 @@ sql.query = {
     restInfo: 'SELECT * FROM Restaurants R INNER JOIN RestaurantStaff RS on R.restaurantID =  RS.restaurantID WHERE RS.uid = $1 LIMIT 1',
     menuInfo: 'SELECT * FROM Food F INNER JOIN Restaurants R on F.restaurantID = R.restaurantID WHERE R.restaurantID = $1',
     // by default category is selected as 'Indian Cusine' by using \'
-    insertFood: 'INSERT INTO Food(foodName, price, category, RestaurantID) Values($1, $2, \'Indian Cuisine\', $3)',
-
-    
+    insertFood: 'INSERT INTO Food(foodName, price, category, restaurantID) Values($1, $2, \'Indian Cuisine\', $3)'
 
 
 }
