+++ App/views/partials/navbar_rest.ejs
@@ -7,6 +7,9 @@
         <li class="nav-item">
             <a class="nav-link" href="/rest_home">Home</a>
         </li>
+        <li class="nav-item">
+            <a class="nav-link" href="/rest_menu">Menu</a>
+        </li>
          <!-- Not Implemented Yet-->
         <li class="nav-item">
           <a class="nav-link" href="#">Logout</a>
